sw: fix cursor backup in SwXText::insertTextPortion()
If the inserted string contains "\r", such as when importing
lp556169-2.rtf before the previous fix, the given cursor has one of its
positions mangled because its nNode will be on the new node, but the
nContent that is restored here is on the old node.

The invalid cursor causes an assert in the subsequent SetPropertyValues.

Change-Id: I73a2e6ecd02ccf75f7b9ffd89d90ae1e3a93ceb2