Bin pointless macro
It is hard to see what the programmer thought the gain of using this
macro was.

It is not any shorter than its expansion. It is also repeated on every
one of the lines in question. If such a macro would be useful,
shouldn't there be some more global place for it, and shouldn't it be
used everywhere then? Why come up with that convention just for this
one file? Etc.

But yeah, I guess the programmer was guarding against mistyping
com.sun.star in some of the lines. As if all those lines would have
been manually typed in character by character anyway.

Change-Id: I2e4354363181ffa55e0a51832e9799f616a7c9f1