gbuild-to-ide: testVS2013Ide .props
this patch add a PropertySheet.props in the windows directory
in this file can put the enviroment variables
for vs2013.

i add in the gbuildparser a variable _jsonsrcdir that
contains the SRCDIR path that it can found in the json file.

with this file we can use enviroment variables of LO.
e.g.: $(SRCDIR), $(BUILDDIR) etc...
in the VS2013 ide.

Change-Id: Ie3158fd31fe247833115d445553f613dcd1a3b1d
Reviewed-on: https://gerrit.libreoffice.org/34749
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: jan iversen <jani@libreoffice.org>