Revert "tdf#106283: Registry settings are not read properly on Windows"
This reverts commit a08b3347a00f888f1b6b3cddf0a72cd6e7750f26.  The problem in
that issue is that the configmgr/source/winreg.cxx code generates .xcu data
where such an extension prop doesn't have an oor:type attribute, which is not
allowed.  On the other hand, it is important that the PropertyNode representing
such an extension prop has a staticType_ of TYPE_ANY, so that later layers or
programmatic calls can store values of different type.