convert VclReferenceBase::mnRefCnt to oslInterlockedCount
Which means that ref-counting VCL widgets is now thread-safe.
Destroying them, however, is not safe without holding the SolarMutex,
which will the subject of a future patch.

According to my tests running 'perf stat make check', this costs us
approx 2% runtime on a non-debug Linux build using gcc.

Costs are quite a bit higher when building with clang, which might need
further investigation.

Change-Id: I737170b74ec79e6b7c7d586ed5be89a73ee83597
Reviewed-on: https://gerrit.libreoffice.org/34948
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>