Simplify and sanitize some OSL_DEBUG_LEVEL > 2 (!) SAL_INFO code
No point in outputting the enum class name when we know it is a value
from that enum that we are outputting.

Fix some random whitespace craziness.

No point in putting that symbolic representation of an enumeration in
double-quotes. It is not a string in any sense.

We can pass an OUString directly in SAL_INFO.

Be consistent, put a double-quote on both sides of the string (font
family name), if at all.

This is a way too verbose commit message for such a trivial change.

Change-Id: I9e3f96856a634dc16f0e24f789c1898d7ec75fd9