tdf#105729 RTF import: \ltrpar should not override \qc from style
This is similar to commit 396ef8cc85b265aa484ab283ab8e0d39fb5aa09f
(tdf#94435 RTF import: \ltrpar should not override \qc, 2015-10-05),
except that here the \qc is inherited from the style, it's not a direct
formatting. The problematic code was added in commit
2638faa2e834c2da4c195224fd88d32c29b3d0cc (writerfilter08ooo330: applied
patch for writerfilter08, 2010-07-28), and it's not really clear to me
what is its purpose, given that the DOC import equivalent in
SwWW8ImplReader::Read_ParaBiDi() doesn't set the paragraph alignment.

Fix the situation by not touching the paragraph alignment for the RTF
case at least.

Change-Id: I2baa2c8c8012d972740da7cf3f710117812859b3
Reviewed-on: https://gerrit.libreoffice.org/35163
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>