vcl: have SolarThreadExecutor use osl::Condition, not oslCondition
Condition is deprecated already, but there is no need for the
SolarThreadExecutor class to use the low-level C-API, when in fact there
is a C++ fascade that calls on this via the C++ abstraction,
osl::Condition.

This will make it much easier to switch to using std::condition_variable
in the future.

Change-Id: I81e8530d42fe1451c4ca5bce998b13de0576bc4e
Reviewed-on: https://gerrit.libreoffice.org/35386
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>
Tested-by: Chris Sherlock <chris.sherlock79@gmail.com>