vcl: have AcquaSalInstance use osl::Condition instead of oslCondition
Condition is deprecated already, but there is no need for the
AcquaSalInstance class to use the low-level C-API, when in fact there
is a C++ fascade that calls on this via the C++ abstraction,
osl::Condition.

This will make it much easier to switch to using std::condition_variable
in the future.

Change-Id: Ic495c4120a59480bf50a8c5b73608874fc4228ea
Reviewed-on: https://gerrit.libreoffice.org/35392
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>