tdf#106606: Don't dispose NumPicBullet in sprm handler.
An image saved in cache once when NS_ooxml::LN_CT_NumPicBullet_pict
is processed, may then be used multiple times (for each NumRule that
requires it) when NS_ooxml::LN_CT_Lvl_lvlPicBulletId is processed
for each of them.

If the image was released here for first processing, subsequent rules
couldn't find the image in cache and failed to create NumberingType::BITMAP
style for the rule.

The image is ultimately released in ListsManager::~ListsManager()
after it is no more needed.

Change-Id: Ib4c351437ba94d5a9d3e2927ccf459ec01f1b15f
Reviewed-on: https://gerrit.libreoffice.org/35591
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>