comphelper::ThreadPool: guard against concurrent shutdown/pushTask
To join a thread, the mutex must be released - another thread
in pushTask() could add a new thread to maWorkers at that point,
which must of course not be deleted.

Avoid the problem by transferring ownership of the to-be-deleted
threads to the calling thread.

(regression from b154ab96c6ba529d7fefd84a2a715e939fd349cd)

Change-Id: I9d4fcfe4cb46a336586b5663934a12d47b2d8ccb