lok sd: Fix crash when searching.
This is a follow-up of 8ac78b401565c5795619fcdc1365867cd3034a1f where the
check for the HasView has been removed.  Turns out that there are conditions
when this really can happen, leading to crashes in the LOK searching.

Unfortunately I did not manage to find a reliable reproducer to create a
unit test :-( - so I suspect this commit might be more a workaround than a
root cause fix.  Would be great to find out the exact conditions leading to
the situation when the EditEngine does not contain the EditView, and evaluate
this fix against that - but that's hard without a reliable reproducer.

A unit test for this fix is missing from the same reason.  The unit test from
8ac78b401565c5795619fcdc1365867cd3034a1f still passes.

Change-Id: I1cca7219817119d27a224b35efb660a84d35b8fa