lok - sc: document size as sum of row heights/col widths in pixel
Grid lines, cursor overlay, row/col headers are all computed by
summing up row heights / col widths converted to pixels.

On the contrary the document size was converted to pixel only at the
end after having summed up heights/widths in twips.

All that lead to have a document height/width greater than the
position of the last row/col, with the scrolling in online going
unplesantly far beyond the last row/column.

This patch change the way the document size is computed, so that the
spreadsheet height/width matches the position of the last row/column.

Moreover it exploits the cache-like structure for row/col positions
introduced in a previous commit.

Change-Id: Ibb2cc6a7b592e359a0b1202dc9bea1dd4c421354
Reviewed-on: https://gerrit.libreoffice.org/40448
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Marco Cecchetti <mrcekets@gmail.com>