ucb: webdav-curl: tdf#106955: Open WebDAV resources on which PROPFIND fails
When PROPFIND fails on a WebDAV resource, its IsDocument property
stays undefined, and so stream creation fails. Proposed solution
is to default to IsDocument=true for all WebDAV documents where
we cannot get the property from server.

Such resources also fail to return their locking options, so
defaulting to server properties. When later locking is attempted
on it, the attempt fails with user notification (a dialog saying
that getting information from server failed). Proposed solution
is to check Content-Disposition header in such resources, and in
case it's attachment, disable lock on this resource. The rationale
for this is that "In a regular HTTP response, the Content-Disposition
response header is a header indicating if the content is expected
to be displayed ... as an attachment, that is downloaded and saved
locally" (see MDN:
https://developer.mozilla.org/en/docs/Web/HTTP/Headers/Content-Disposition

Also, Content::getProperties wasn't ready for PROPFIND returning
empty result.

[ port of commit 1eea49feb1585adfa07a140c6e1d0bddc4984e03 ]

Change-Id: If0b4c10ef7b7b108a8779c773c65e25973d32b46
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123504
Tested-by: Michael Stahl <michael.stahl@allotropia.de>
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>