Fix include and library paths for at least me
Despite that the HKLM/SOFTWARE/Microsoft/Microsoft SDKs/Windows/
v10.0/ProductVersion value is 10.0.15063, the headers and libraries we
want are in the subdirectories called 10.0.14393.0, not 10.0.15063.0.

Not sure if this is just because of the collection of Visual Studio
versions I happen to have installed, and their order of installation,
or always true.

Instead of all the convoluted logic in configure.ac, it would make
much more sense to just use the relevant vcvars64.bat or vcvars32.bat
file and pick out the INCLUDE and LIB environment variables it sets,
and mangle that. We already have a (misleadingly named) function to
look for variables set by a .bat file, win_get_env_from_vsvars32bat().

Change-Id: I647ec1d0924ba4c743e4c16d2fa0a05014e0b2e3