tdf#107211 sw: ODF import: add a horrible hack to import "Default"
... character style for hyperlinks and index-entry-templates.

The problem is that the "Default" character style doesn't actually
exist, it is just displayed in the UI so you can remove a character
style that is set.  So for the most part there is no need to store
"Default" in ODF files, except that for hyperlinks the default isn't
"Default" but "Internet Link"/"Visited Internet Link".

Hence it was not really a good idea to rename "Default" to "Default
Style", because when importing existing documents the
text:style-name="Default" is not found, and for new documents
text:style-name="Default_20_Style" is not found either because
there is no style:style element that sets up the mapping between the
encoded style-name and the unencoded display-style-name.

Add some hack in SwXMLBodyContext_Impl::SwXMLBodyContext_Impl(), when
all style:style elements have been read (so we don't override what's
in the document) to create the mapping for "Default", "Default_20_Style"
and whatever the translation of that is.

(regression from 15ab72e1d32063d1dd8f1a97febd774cbfac8959)

Change-Id: I8a80847571f194204c5df267cb48e826a82328b2