tdf#107126 sw: fix layout crash with section in footnote
SwContentFrame::WouldFit_() contains a hack to temporarily reparent a
SwTextFrame.  In the bugdoc, there is a SwTextFrame below a
SwSectionFrame below a SwFootnoteFrame.  The reparenting ignores the
SwSectionFrame so the result is a SwTextFrame below SwFootnoteFrame, but
it still has its mbInfSct set, hence crashes with a null pointer.

If the SwTextFrame is permanently moved later on, in
SwFlowFrame::MoveBwd() line 2450 a new SwSectionFrame is created.

Change-Id: I45a7ab793b4459e551bd11b7fb83dedc58a6c8da